fix: add all KPIs to "All" dashboard on upgrading
All existing KPIs (even inactive ones) are added to the newly created
"All" dashboard on upgrade.

On downgrade, all KPIs are unregistered from the "All" dashboard and
then the "All" dashboard is deleted. If there are multiple active "All"
dashboards, all of them are deleted and their KPIs are unregistered.

Note: If a KPI has been added to a dashboard prior to upgrading (not
possible if upgrading from last stable version), they will still be
added to the newly created All dashboard.