feat(cli): add command to import KPIs from a JSON
The `kpi-import` command reads a JSON file given as input and adds all
the KPIs in it. The JSON must have a list of KPIs in the specified
format (similar to JSON of the API endpoint - some field names may
differ).

TODO: some of the code is repeated from the KPI endpoint.

If a KPI addition fails, it's logged and the script moves on to the next
one.