fix(alert-digest): avoid "Any" in email dropdown
"Any" was added to every email and added by default to the dropdown.
This was needed because Slack alerts which do not have emails would not
show up when an email filter was used (even if it was empty).

With the new filters architecture, this isn't needed anymore. The email
filter is removed when no emails are selected.