Fix status page containers when airbyte is disabled (#629)
* fix(settings): parse booleans correctly from env

Added a helper functions _make_bool to parse truthy values as True.

Used it for all variables which are booleans. Removed extra code that
was used to parse booleans before this change.

* fix(status): do not show airbyte containers status when its disabled

Split CONTAINERS constant into two - CG and airbyte containers. Only CG
containers are considered when airbyte is disabled.

Also formatted with black.

* fix(integration-utils): remove unnecessary truthy check

These env vars are converted to bool in #629, so this check can be a
simple boolean check.