fix(alerts): change calc was wrong in Overall KPI for daily KPIs
In the change calculation for daily KPIs, overall series percentage
change and message was always shown as "-"

This was because `series_type` for such entries was NULL while
`anomaly_type` was set to overall. Since the percent change logic was
only comparing `series_type`, it was unable to find the last data point
for those entries.