fix: remove subtraction of 1 day in anomaly end date
This was needed because the last date of data in DB was different from
the anomaly end date, since the data loader using `< end_date` for
query.

This behaviour was changed in #661 to use `<= end_date`. Now, the
subtraction is unnecessary and causes alerts to be sent for one day
older anomalies too.