fix(alerts): use Any type for alert_channel_conf since it's a dict
The type was assumed to be `Optional[list]` because it was used as a
list in the alert dashboard template. It appears that the type is
different for email and slack. It's a dict containing `email` as a key
and the list of emails as value.

Used type `Any` to avoid type issues with this field.