chore(ci): add step to check if migrations have diverged
The script prints a message with the IDs of the two heads if migrations
have diverged. Also returns an appropriate error code for CI.

Added this as the last step in CI to check on every PR.

Also removed overridden flake8 arguments in the linting step. We have a
`.flake8` config in the project root for these args.