refactor(kpi): optimize kpi list, paginate dashboards too
Previously, `get_mapper_obj_by_dashboard_ids` would get the list of all
KPIs for the given dashboard_id, the query would use the entire list
(could be 500+ in some cases) embedded in the query itself. After this
query is done, pagination would be applied which doesn't provide much
benefits in terms of DB performance.

Refactored it to use DashboardKpiMapper in the query itself.
- the filter ensures that the entire list of KPIs is not retreived at
  any point
- pagination added in previous commit ensures performance

The `dashboards` field returned by the API endpoint was not used in the
frontend, so it was removed. This meant that a lot of unnecessary code
could be removed from the endpoint. Note: there is another end point to
get the list of dashboards, so this is not necessary.

The logic of associating dashboards with KPIs was refactored:
- Created function `kpi_dashboard_mapper_dict` in `dashboard_controller`
  that creates a dict that can be used to get the list of dashboards for
  given KPIs. This avoids calling multiple of `get_mapper_obj...`
  functions just for this purpose in the view.
- The function is type safe even when the return type depends on a
  parameter. Used `typing.overload` for this.

BREAKING CHANGE: removed `dashboards` fields from GET `/api/kpi`
endpoint.