feat(gitpod): add workers, scheduler, doc comments
Also change GitHub prebuilds config to not add a check to every PR since
it does not actually check anything. Made it add a comment instead.

Also moved `unset PGHOSTADDR` to a `before` block so that it happens
before every init and before starting a workspace. It's set directly in
the `.bashrc` now.

Also added Redis' temp server `dump.rdb` to `.gitignore`.