feat(dashboard-view): add sorting on 3 keys for dashboard list
Used globals to store the default and mapping in one place - adding a
new sort require change to one place only.

Added `sort_by` arg in the list endpoint. Validation is performed on it
too. The sorting is done by `get_dashboard_list` in the controller
instead of directly having it in the view.

Added missing docstring to `get_dashboard_list`.