feat(api): allow multiple values for filter endpoints
Uses getlist [1] to take all values of the parameter instead of just the
first one. The query is changed to use `in_` instead of `==`.

Type conversions are applied to each element in the list as appropriate.
The `type` parameter of `getlist` is not used since it ignores values
that cannot be parsed to that type - we want to show an explicit error
instead.

[1]: https://werkzeug.palletsprojects.com/en/2.1.x/datastructures/#werkzeug.datastructures.MultiDict.getlist