feat(alerts): add top relevant_subdims to AnomalyPointFormatted
Had to write a new helper `extract_dim_value_from_series_type` in
`string_helpers` that takes the series_type string and extracts
dimension and value from it.

Had to store `series_type_original` in AnomalyPoint (which means it's
stored in the TriggeredAlerts table) as the final `series_type` stored
was the user string. This will be `None` for all older entries, which
means subdim data may not be shown for reports with older data.