fix(anomaly): convert y values to float before saving
When the `y` values are integers, anomaly fails while saving with the
error `sqlalchemy.exc.ProgrammingError: (psycopg2.ProgrammingError)
can't adapt type 'numpy.int64'`.

This is noticeable either with KPIs that have their metric column as an
integer type or with Data Quality count.

The commit fixes this by always casting the `y` values to `float`.