fix(anomaly): ensure correct data types
The empty prediction dataframe is now given the correct types
immediately after creation. This ensures that any data appended to the
dataframe will have the correct type. Note: if the columns or the
dataframe itself is overwritten, this will not hold and the types may
change.

Also uses a simpler method to get the last row of the prediction
dataframe which prevents loss of types. The earlier version would erase
all the metadata and make the dtypes `object`.

These 2 changes ensure that the `y` column is always a `float` and the
dataframe correctly identifies it as such. Note: the columns can still
be overriden by some other code and this cannot be ensured without
writing tests.