fix(datasource): use case-insensitive check for LIMIT
Before applying our own `LIMIT 1`, we check if the query already has a
`LIMIT` clause by a simple substring search. This had two issues:
- It did not work when `LIMIT` was in uppercase
- It could match "limit" in a column or table name

Changed this to use a case-insensitive regex search which also looks for
whitespace around `LIMIT`. Since "limit" as a column or table name will
need to enclosed in double quotes, this regex will not match them.