fix(alerts): consider subdims inside relevant only if enabled
A bug existed that appeared in this case:
- An alert is set up with subdims disabled and set to digest/report
- For the same day, both overall and subdim anomalies are present
    - (same hour in case of hourly alerts)

The bug:
The digest/report sent on that day included subdim anomalies for that
KPI in the "Top sub-dimensional anomalies" section even though subdims
was disabled.

The cause:
The subdims inside of relevant_subdims were being considered even if
include_subdims was set to false. This did not affect subdims that were
not associated with an overall anomaly.

The fix:
Add the include_subdims condition to subdims inside relevant_subdims
too.